requirements.txt
config==0.4.2
dj-database-url==0.5.0
Django==1.11.5
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-tinymce==2.8.0
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
whitenoise==4.1.2
